Cache parts of the video memory address.
Greatly simplifies the code paths in the hot part of the video renderer,
but doesn't seem to have the speed impact I'd have hoped it would have.
Checking in despite a few % slowdown (in some runs) as I morally feel it
is a better approach and hopefully opens the door to higher level
optimizations higher up the call stack.