Make bus timings stable
Previously the odd/even-ness of the cycle was determined by the "cycles"
counter which was added-to on cpu.execute(). That meant an odd-numbered
number of cycles would inadvertantly toggle the oddness of the bus
cycle, which made Frogman's protection system fail in different random
places based on the browser's performance (via time-scaled calls to
cpu.execute() in requestAnimationFrame).

Fixes #60.