Implement an embed mode for use within an iframe
Add a new URL parameter of 'embed', which adjusts the entries in the top
menu to make the application easier to embed in third-party pages using
an iframe.

Specifically, the 'Discs', 'Cassettes', 'Share' and 'Contact' drop-downs
are hidden: disc and cassette selection as it would be expected that an
embedded instance would be pre-loaded with a disk or tape and share and
contact as they are rather confusing if the app is embedded in a third-
party site.  The 'About' entry is also changed to 'About jsbeeb', again
for clarity.

Also set the href of the navbar header to point to
http://bbc.godbolt.org/ so that interested users can easily click through
to the main jsbeeb page.