Swapped key codes for mac users.
or Mac users, the key codes for APOSTROPHE and BACK_QUOTE seem to be
the wrong way around compared what jsbeeb expects.  This makes it
almost impossible to play games where the BBC.COLON_STAR key is used
for “up”.  This fix checks to see if the user has a Mac keyboard and
swaps the key codes around so that they match what the user expects.