Update utils.js
Set the `BACK_QUOTE` and `APOSTROPHE` keys to the values that jsbeeb expects for Mac users.  I also set the `HASH` keycode to 223 to prevent duplicate mappings.  (The `HASH` key does not exist on modern Mac keyboards and instead we have to use ALT+3.  It makes working Python painful.)

The stanza of code was also moved upstream to make sure that all choices of keyboard layout see the same change of mapping.