Fix disappearance of carrier tone if motor is toggled. (#254)
Fixes #241

The problem is triggered by toggling the motor off/on at an inopportune time. Turning the motor off cancels any in-progress UEF blocks that schedule a finish time.

In the case of Atic Atac, blipping the motor seems to be canceling (effectively truncating) the carrier tone blocks between the data blocks. Normally this might not be a problem but Atic Atac has a custom loader that gets into timing trouble. (I think it's using MOS to load blocks with a bad data checksum, and snooping the serial load space via listening to the serial buffer full event or something like that.)