Allow tokenising longer input lines
We used the tokenisation code in the BASIC ROM, which calls a subroutine
to copy down the tail of the line after substituting a token, using a loop
indexed on Y.  This limits the untokenised tail to 255 bytes (plus a
terminating CR) and makes tokenising O(n²) for a line with a lot of tokens.

Intercept the instruction which calls this subroutine, copy out the
newly processed bytes, adjust the zero page pointer and resume executing
at the instruction after the subroutine call.  This allows tokenising
arbitrarily long input provided the result fits within the BASIC line
length limit, and is faster.