Fix cub monitor/screen size bug
Seemed the document is now not quite ready when we first are called
and that meants the cubOrigHeight was wrong (as we were measuring
the parent that contained an image). This fix uses the actual parts
that are important - the height/width attribute of the image holder.

I'm not 100% sure why really, but empirically this fixes the issue
which only showed up in "--node-env production" builds.