Handle cases where the counter went super-negative
Ironically during NegativeCharge's awesome ports of various tunes the
step counter for the noise channel would go super-negative, which
meant it would "squeaaaaaak" until it clawed its way back positive.

The real solution is to output at 250kHz and downsample, and get
rid of all the complexity here. I'll do that next.

Thanks to @richtw1, @scarybeasts and @NegativeCharge and the rest of
the bitshifters crew.

Closes #334