In 'embed' mode, make screen as large as possible
Currently there is always a 100 pixel margin added on the left and right
sides of the screen and the same (minus the header bar top and bottom
margins) added at the bottom.  When jsbeeb is embedded in an <iframe>
and the 'embed' parameter is present, remove these margins so that the
screen uses as much of the available space as possible.

This also required switching from using the .height() of the header bar
to the .outerHeight() for accurately calculating the remaining vertical
space.  Reduce the margin size value for the bottom to 68 pixels to
compensate for this change.

Also update the README entry for 'embed' to include this behaviour (and
remove the reference to the fixed iframe size as this is no-longer valid
since screen autosizing was implemented).