Make the page background transparent in embed mode
Override the page background colour when in 'embed' mode, setting it to
transparent so that the background of the parent page is shown instead.

Setting very different background colours revealed a gap between the
bottom of the screen canvas element and the middle of the bottom frame
of the monitor.  To resolve this, updated this section of the image to
set it to black (which was the original colour of this area before it
was converted from jpeg to transparent png).

Also, while editing the monitor frame image set the black part
underneath the monitor to transparent as this gives a neater result with
non-black page backgrounds (such as the default grey).