A collection of (mostly) embed mode improvements (#363)
* Remove unused 'embed' code for About menu entry

This is both redundant (as the menu entry includes jsbeeb as a suffix
anyway) and also unnecessary as the 'More' menu is hidden in embed mode
so the change cannot be seen.

* Move audio warning to float on top of screen

This means that it doesn't move the screen downwards when shown and then
back up again after being hidden.

* In 'embed' mode, make screen as large as possible

Currently there is always a 100 pixel margin added on the left and right
sides of the screen and the same (minus the header bar top and bottom
margins) added at the bottom.  When jsbeeb is embedded in an <iframe>
and the 'embed' parameter is present, remove these margins so that the
screen uses as much of the available space as possible.

This also required switching from using the .height() of the header bar
to the .outerHeight() for accurately calculating the remaining vertical
space.  Reduce the margin size value for the bottom to 68 pixels to
compensate for this change.

Also update the README entry for 'embed' to include this behaviour (and
remove the reference to the fixed iframe size as this is no-longer valid
since screen autosizing was implemented).

* Make the page background transparent in embed mode

Override the page background colour when in 'embed' mode, setting it to
transparent so that the background of the parent page is shown instead.

Setting very different background colours revealed a gap between the
bottom of the screen canvas element and the middle of the bottom frame
of the monitor.  To resolve this, updated this section of the image to
set it to black (which was the original colour of this area before it
was converted from jpeg to transparent png).

Also, while editing the monitor frame image set the black part
underneath the monitor to transparent as this gives a neater result with
non-black page backgrounds (such as the default grey).