Make legacy model handling case insensitive too (#364)
Handling of the current model names is case insensitive, e.g.
model=master is valid as well as model=Master.  mapLegacyModels is
however case sensitive so model=B works but model=b causes an uncaught
TypeError.

Resolve this inconsistency by making mapLegacyModels case insensitive.
Also took the opportunity to move handling of the legacy 'B' model to
just be another alias of 'BBC B with DFS 0.9' and corrected a comment.